third_party/blink/web_tests/platform/generic/fast/forms/hidden/hidden-setrangetext-expected.txt

Test setRangeText() method is not available in hidden inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Running tests on input with attributes: {"type":"hidden"}

element.value = '0123456789XYZ'
PASS element.setRangeText('ABC', 0, 0) threw exception InvalidStateError: Failed to execute 'setRangeText' on 'HTMLInputElement': The input element's type ('hidden') does not support selection..
PASS element.value is "0123456789XYZ"
PASS successfullyParsed is true

TEST COMPLETE
